arm64: dts: rk3399-pinephone-pro: Add flash and fix leds
The PinePhone Pro has a similar device to the SGM3140 used in the OG
PinePhone. This commit adds the corresponding device-tree node so the
flash/torch can work on this device.

Moreover, existing LEDs use different names and functions than what
we're used to, change them all to the same names we use on the OG
PinePhone.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@gmail.com>

Gbp-Pq: Topic device-tree
Gbp-Pq: Name 0114-arm64-dts-rk3399-pinephone-pro-Add-flash-and-fix-led.patch