power: supply: ip5xxx: Add support for POWER_SUPPLY_PROP_CHARGE_BEHAVIOUR
This property is better for preventing charging, then the status property,
because reading status property should return the actual status, and not
whether the charging is currently inhibited or not.

System may have charging enabled, but the battery may still be discharging
for a host of reasons.

Signed-off-by: Ondrej Jirman <megi@xff.cz>

Gbp-Pq: Topic keyboard
Gbp-Pq: Name 0083-power-supply-ip5xxx-Add-support-for-POWER_SUPPLY_PRO.patch