drm: rockchip: Don't require MIPI DSI device when it's used for ISP
On RK3399 one MIPI DSI device can be alternatively used with the ISP1,
to provide RX DPHY. When this is the case (ISP1 is enabled in device
tree), probe success of DRM is tied to probe success of ISP1 connected
camera sensor. This can fail if the user is able to killswitch the camera
power, like on Pinephone Pro.

Detect use of MIPI DSI controller by ISP, and don't include it in
component match list in that case.

Signed-off-by: Ondrej Jirman <megi@xff.cz>

Gbp-Pq: Topic display
Gbp-Pq: Name 0017-drm-rockchip-Don-t-require-MIPI-DSI-device-when-it-s.patch