usb: serial: option: add 'reset_resume' callback for WWAN devices
This avoids the issue where the PPP modem does not reset correctly on resume,
causing issues with incoming calls on suspend.

Noted here: https://github.com/the-modem-distro/pinephone_modem_sdk/issues/97

This is part of a larger patch:
https://github.com/Quectel-Community/meta-quectel-community/commit/24adef2128f79b4d31490634438a917b3df462e2#diff-45dc5960ca2f1ef57139fe751587b3b043de9cca5864e2801210c00e75bd0780

However the rest of the patch is not needed/already upstreamed.

Gbp-Pq: Topic common
Gbp-Pq: Name 0007-usb-serial-option-add-reset-resume-callback-for-WWAN.patch