drm: panel: hx8394: Fix mode to have refresh rate of 60 Hz
VSYNC settings did not match the init sequence values. Fix that first.

Now we need to generate exact clock frequency from some of the available
PLLs by a simple division:

  GPLL = 594 MHz
  VPLL = unused, can be set to anything
  CPLL = 800 MHz

Needed clock is 60 * vtotal * htotal = 60 * 1463 * ~844 = 74086320

So no we have an optimization task. We need clock frequency that can
be generated from 594 or 800 MHz by integer division and we can adjust
the needed frequency a little by adjusting hsync timings.

  594 / 74 = 8.027
  800 / 74 = 10.811

So GPLL is a better candidate because it will need smaller adjustment
of HSYNC timings. Exact frequency that can be generated is:

  594 / 8 = 74.25

Which would require htotal to be 74 250 000 / (60 * vtotal) = 846
So let's set horizontal timings to 720 40 46 40.

Signed-off-by: Ondrej Jirman <megi@xff.cz>

Gbp-Pq: Topic display
Gbp-Pq: Name 0021-drm-panel-hx8394-Fix-mode-to-have-refresh-rate-of-60.patch