arm64: dts: rk3399-pinephone-pro: reduce thermal limits
While this device uses the rk3399 it is also enclosed in a tight package
and cooled through the screen and back case. The default rk3399 thermal
limits can result in a burnt screen.

These lower limits have resulted in the existing burn not expanding and
will hopefully result in future devices not experiencing the issue.

Signed-off-by: Jarrah Gosbell <kernel@undef.tools>

Gbp-Pq: Topic device-tree
Gbp-Pq: Name 0132-arm64-dts-rockchip-reduce-thermal-limits-on-PinePhon.patch