input: pinephone-keyboard: Wait a bit after enabling vbus
In case the MCU was browned out, we want to wait a bit after enabling
VBUS and before probing the MCU over I2C, for the MCU to initialize.
Otherwise it may be possible for the MCU to fail to respond during probe
and for the driver to fail initialization.

Signed-off-by: Ondrej Jirman <megi@xff.cz>

Gbp-Pq: Topic keyboard
Gbp-Pq: Name 0085-input-pinephone-keyboard-Wait-a-bit-after-enabling-v.patch