d/patches: fix bug reporting 0% battery charge
The threshold for the "low battery" condition is set too low, causing
issues where the battery driver reports 0% remaining charge while it's
actually way higher. This in turn causes g-s-d to auto-shutdown the
device even when active. Let's revert to the previous threshold value so
we don't run into this problem.